+++ ai/app/api/dto/additional_infos_request.py
